requirements.txt
# ===========================================
# MMarra Data Hub - Dependências
# Versão 2.0 - Fevereiro 2026
# ===========================================

# ============ CORE ============
python-dotenv>=1.0.0          # Variáveis de ambiente
requests>=2.31.0              # HTTP requests
pandas>=2.0.0                 # Manipulação de dados
pyarrow>=14.0.0               # Formato Parquet
numpy>=1.24.0                 # Computação numérica

# ============ AZURE ============
azure-storage-file-datalake>=12.14.0  # Data Lake Gen2
azure-identity>=1.15.0                 # Autenticação Azure

# ============ TEMPLATES / RELATÓRIOS ============
jinja2>=3.1.0                 # Templates HTML

# ============ MACHINE LEARNING ============
scikit-learn>=1.3.0           # ML geral (clustering, anomalias)
prophet>=1.1.0                # Previsão de séries temporais

# ============ LLM / LANGCHAIN ============
langchain>=0.1.0              # Framework LLM
langchain-core>=0.1.0         # Core do LangChain
langchain-groq>=0.0.1         # Integração Groq

# ============ MCP SERVER ============
mcp>=0.1.0                    # Model Context Protocol

# ============ DESENVOLVIMENTO ============
pytest>=7.4.0                 # Testes
black>=23.0.0                 # Formatação de código
isort>=5.12.0                 # Ordenação de imports
flake8>=6.1.0                 # Linting

# ============ OPCIONAIS ============
# Descomente conforme necessidade:

# plotly>=5.18.0              # Gráficos interativos
# streamlit>=1.29.0           # Dashboard web
# schedule>=1.2.0             # Agendamento de tarefas
# fastapi>=0.104.0            # API REST
# uvicorn>=0.24.0             # Servidor ASGI
